# Share the TritonTF model across instances on same device