@CoderHam 2021-03-17
While the performance look similar, how is the memory consumption affected if we run 2 model instances on a single GPU?