@tanmayv25 2021-03-18
This PR will allow us to set higher instance_count for TensorFlow as the sessions will be shared. The performance with the same instance count is expected to be within similar range as there will be contention over the resources. But having a parameter to control how many instances to share a session will be a useful tuning knob. This should tie in well with `TF_USE_PER_SESSION_THREADS` knob exposed in PR #15 to allay some of the CPU side contention with per session thread pools.